[wperf-lib] Update wperf_sample API to support 0.1s sampling duration accuracy
See merge request Linaro/WindowsPerf/windowsperf!286

=== ChangeLog ===

* [wperf-lib] Update wperf_sample API to support 0.1s sampling duration accuracy

The wperf_sample API currently supports sampling duration accuracy of 1s. With
!278, wperf app now supports sampling duration accuracy of 0.1s. This patch
aligns the wperf_sample implementation with the wperf app in this aspect.

+++ wperf-lib/wperf-lib.h
@@ -452,8 +452,8 @@ typedef struct _SAMPLE_CONF
     uint32_t *intervals;
     /// Set this to true to get short symbol names, false to get long symbol names.
     bool display_short;
-    /// Sampling duration in second.
-    uint64_t duration;
+    /// Sampling duration (in seconds, accurary of 0.1 sec).
+    double duration;
     /// Set this to true if kernel mode should be included, false if not.
     bool kernel_mode;
     /// Set this to true if funtions should be annotated, false if not.
